Implement #4

+++ python/localz/__main__.py
@@ -8,11 +8,13 @@ import logging
 import argparse
 import tempfile
 import traceback
+import itertools
 import contextlib
 
 from rez.config import config
 from rez.resolved_context import ResolvedContext
 from rez.packages_ import Package, Variant, iter_packages
+from rez.exceptions import PackageFamilyNotFoundError
 from rez import package_copy, __version__ as rez_version
 
 try:
@@ -33,7 +35,7 @@ except ImportError:
     version = "dev"
 
 description = """\
-Localise packages from one location to another, to improve
+Localize packages from one location to another, to improve
 performance or facilitate remote collaboration on software
 or content.
 """
@@ -41,13 +43,13 @@ or content.
 epilog = """\
 examples:
   Latest available version of Maya
-  $ rez env localz -- localise maya
+  $ rez env localz -- localize maya
 
   Version of Maya, required by Alita
-  $ rez env localz -- localise maya --requires alita
+  $ rez env localz -- localize maya --requires alita
 
   Series of requests, all compatible with each other
-  $ rez env localz -- localise six-1 maya-2018 python-3.7
+  $ rez env localz -- localize six-1 maya-2018 python-3.7
 
 """
 
@@ -61,19 +63,28 @@ parser.add_argument("request", nargs="*", help=(
 parser.add_argument("--version", action="store_true", help=(
     "Print version and exit"))
 parser.add_argument("--requires", nargs="+", default=[], metavar="PKG", help=(
-    "Localise request, fulfilling these requirements"))
+    "Localize request, fulfilling these requirements"))
 parser.add_argument("--all-variants", action="store_true", help=(
     "Copy not just the resolved variant, but all of them"))
 parser.add_argument("--paths", nargs="+", metavar="PATH", help=(
-    "Package search path (ignores --no-local if set)"))
+    "Override package search path"))
 parser.add_argument("-f", "--force", action="store_true", help=(
     "Copy package even if it isn't relocatable (use at your own risk)"))
-parser.add_argument("-v", "--verbose", action="count")
+parser.add_argument("-v", "--verbose", default=0, action="count")
+parser.add_argument("--full", action="store_true", help=(
+    "Localize requests and requirements of requests. "
+    "Use this to create a fully localized context"))
 
 
-log = logging.getLogger(__name__)
-log.setLevel(logging.DEBUG)
 opts = parser.parse_args()
+log = logging.getLogger(__name__)
+
+logginglevel = {
+    0: logging.INFO,
+    1: logging.DEBUG,
+    2: logging.DEBUG,
+}.get(opts.verbose, logging.INFO)
+log.setLevel(logginglevel)
 
 
 def tell(msg, newlines=1):
@@ -82,12 +93,21 @@ def tell(msg, newlines=1):
 
     sys.stdout.write("%s%s" % (msg, "\n" * newlines))
 
+    if sys.stdout.isatty():
+        # Give the user some time to react
+        time.sleep(0.01)
+
 
 def warn(msg, newlines=1):
     if log.level > logging.WARNING:
         return
 
-    sys.stderr.write("%s%s" % (msg, "\n" * newlines))
+    sys.stderr.write("WARNING: %s%s" % (msg, "\n" * newlines))
+
+
+def abort(msg):
+    sys.stderr.write("ABORTED: %s\n" % msg)
+    exit(1)
 
 
 if opts.version:
@@ -101,12 +121,17 @@ if not opts.request:
 
 
 def cleanup():
+    if not os.path.exists(tempdir):
+        return
+
     try:
+        sys.stdout.write("Cleaning up temporary files.. ")
         shutil.rmtree(tempdir)
+        sys.stdout.write("ok\n")
     except OSError as e:
         # It may already have been cleaned-up
         if e.errno not in (errno.ENOENT, errno.EEXIST, errno.ENOTDIR):
-            print("errno: %s" % e.errno)
+            sys.stdout.write("fail\n")
             traceback.print_exc()
 
 
@@ -137,26 +162,62 @@ def ask(msg):
         return False
 
 
+class Animation(object):
+    frames = itertools.cycle(r"\|/-")
+
+    def __init__(self, template="{frame}", count=10):
+        self._template = template
+        self._length = len(template)
+        self._count = count
+
+    def __next__(self):
+        self.tell(next(self.frames))
+
+    def step(self):
+        next(self)
+
+    def tell(self, frame):
+        message = self._template.format(frame=frame)
+        message = "\r%s" % message
+
+        sys.stdout.write(message)
+        sys.stdout.flush()
+
+    def finish(self):
+        self.tell("")
+
+
 @contextlib.contextmanager
-def stage(msg, timing=True):
-    verbose = log.level < logging.INFO
-    tell(msg, 0)
+def stage(msg, count=0):
     t0 = time.time()
 
-    try:
-        yield
-    except Exception:
-        if not verbose:
-            tell("fail")
-        raise
-    else:
-        if not verbose:
-            return
+    if count:
+        msg = "%s {frame}" % msg
 
-        if timing:
-            tell("ok - %.2fs" % (time.time() - t0))
+    bar = Animation(msg, count)
+
+    try:
+        next(bar)
+        yield bar
+
+    except Exception as e:
+        bar.finish()
+        tell("fail")
+
+        if opts.verbose == 1:
+            tell(e)
+
+        elif opts.verbose > 1:
+            traceback.print_exc()
+
         else:
-            tell("ok")
+            tell("Pass --verbose for more details")
+
+        exit(1)
+
+    else:
+        bar.finish()
+        tell("ok - %.2fs" % (time.time() - t0))
 
 
 # Keep it tidy
@@ -169,26 +230,63 @@ tell("Using %s-%s" % (__project__, rez_version))
 
 # Find local packages path
 variants = list()
-paths = opts.paths or config.nonlocal_packages_path
+nonlocal_packages_path = opts.paths or config.nonlocal_packages_path
+localized_packages_path = config.local_packages_path
 
+tell("Packages requested: %s" % " ".join(opts.request))
+tell("Packages will be localized to %s" % localized_packages_path)
 tell("Packages are discovered from these paths:")
-for path in paths:
+for path in nonlocal_packages_path:
     tell("  %s" % path, 1)
 
-with stage("Resolving requested packages.. "):
-    context = ResolvedContext(opts.request + opts.requires)
+with stage("Resolving requested packages.."):
+    try:
+        context = ResolvedContext(opts.request + opts.requires)
+
+    # Handle common errors here
+    # The rest goes to the handler in stage()
+    except PackageFamilyNotFoundError as e:
+        sys.stdout.write("fail\n")
+
+        # Ideally wouldn't have to parse the string-output of
+        # this exception, but there isn't anything else we can do.
+        try:
+            _, package = str(e).split(":", 1)
+            package, _ = package.split("(", 1)
+            package = package.strip()
+
+        except Exception as e:
+            # In case the string value changes
+            abort(str(e))
+
+        else:
+            abort(
+                "Package '%s' wasn't found in any of your "
+                "non-local package paths" % package
+            )
 
     # Sort out relevant packages
     for variant in context.resolved_packages:
-        if variant.name not in opts.request:
-            continue
+
+        if not opts.full:
+            # Include only requested packages
+            if variant.name not in opts.request:
+                continue
 
         variants += [variant]
 
 
+count = len(variants)
 copied = list()
-with stage("Preparing packages.. "):
+skipped = list()
+with stage("Preparing packages..", count) as bar:
     for var in variants:
+
+        # Don't want to localise already-localised packages
+        if var.is_local:
+            skipped += [var.resource]
+            continue
+
         result = package_copy.copy_package(
             package=var.parent,
 
@@ -199,12 +297,14 @@ with stage("Preparing packages.. "):
             shallow=False,
             follow_symlinks=True,
 
-            # Make localised packages as similar
+            # Make localized packages as similar
             # to their original as possible
             keep_timestamp=True,
 
             force=opts.force,
-            verbose=opts.verbose,
+
+            # Only for emergencies
+            verbose=opts.verbose > 2,
         )
 
         for source, destination in result["copied"]:
@@ -215,8 +315,10 @@ with stage("Preparing packages.. "):
         # to ensure it doesn't proceed under false pretenses.
         assert not result["skipped"], (tempdir, result["skipped"])
 
+        bar.step()
 
-with stage("Determining relocatability.. "):
+
+with stage("Determining relocatability.."):
     def is_relocatable(pkg):
         if pkg.relocatable is None:
             return config.default_relocatable
@@ -237,50 +339,21 @@ if unrelocatable:
     shutil.rmtree(tempdir)
     exit(1)
 
-skipped = list()
-local_packages_path = config.local_packages_path
-with stage("Determining existing packages.."):
-    for variant in copied[:]:
-        it = iter_packages(variant.name,
-                           range_=str(variant.version),
-                           paths=[local_packages_path])
-
-        for package in it:
-            for existing in package.iter_variants():
-                variant = Variant(variant)
-
-                if existing.name != variant.name:
-                    continue
-
-                if existing.version != variant.version:
-                    continue
-
-                if existing.index != variant.index:
-                    continue
-
-                # Already exists!
-                skipped += [variant]
-
-    for existing in skipped:
-        copied.remove(variant)
-
 
 if not copied:
-    tell("All requested packages were already localised")
+    tell("All requested packages were already localized")
     shutil.rmtree(tempdir)
     exit(0)
 
 
-tell("The following NEW packages will be localised:")
+tell("The following NEW packages will be localized:")
 for variant in copied:
     tell("  %s-%s" % (variant.name, variant.version))
 
 if skipped:
-    tell("The following packages will be SKIPPED:")
+    tell("The following packages were already localized:")
     for variant in skipped:
         tell("  %s-%s" % (variant.name, variant.version))
-
-tell("Packages will be localized to %s" % local_packages_path)
 
 size = sum(
     os.path.getsize(os.path.join(dirpath, filename))
@@ -303,7 +376,7 @@ for variant in copied:
     pkg = Package(variant.parent)
     result = package_copy.copy_package(
         package=pkg,
-        dest_repository=local_packages_path,
+        dest_repository=localized_packages_path,
         shallow=False,
         follow_symlinks=True,
         keep_timestamp=True,
@@ -312,7 +385,7 @@ for variant in copied:
     )
 
     if result["skipped"]:
-        print("These were already localised")
+        print("These were already localized")
 
 tell("Success")
